ft_trans/backend/users/requirements.txt
psycopg2-binary>=2.9,<3.0
channels==4.0.0
Pillow==10.0.0
Django==4.2.6
PyJWT==2.7.0